[BOD-395] Remove text & button props & allow children elements
