+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agroptima-design-system",
-  "version": "0.22.3",
+  "version": "0.22.4",
   "scripts": {
     "dev": "npm run storybook",
     "storybook": "storybook dev -p 6006 --ci",
