+++ src/stories/Changelog.mdx
@@ -4,6 +4,12 @@ import { Meta } from "@storybook/blocks";
 
 # Changelog
 
+# 0.22.4
+
+BREAKING CHANGES
+
+* EmptyState text and button props have been removed and children elements can be passed to the component.
+
 # 0.22.3
 
 * Add Delivery Note icon
